moved gen24 to fronius directory
