--- io.openems.edge.fronius.gen24.neu/.gitignore
@@ -1,2 +1,0 @@
-/bin_test/
-/generated/
