--- io.openems.edge.fronius.gen24.neu/bnd.bnd
@@ -1,20 +1,0 @@
-Bundle-Name: OpenEMS Edge Gen24neu
-Bundle-Vendor: FENECON GmbH; OpenEMS Association e.V.
-Bundle-License: https://opensource.org/licenses/EPL-2.0
-Bundle-Version: 1.0.0.${tstamp}
-
--buildpath: \
-	${buildpath},\
-	io.openems.common,\
-	io.openems.edge.battery.api,\
-	io.openems.edge.batteryinverter.api,\
-	io.openems.edge.bridge.modbus,\
-	io.openems.edge.common,\
-	io.openems.edge.ess.api,\
-	io.openems.edge.meter.api,\
-	io.openems.edge.pvinverter.api,\
-	io.openems.edge.timedata.api,\
-	io.openems.j2mod,\
-
--testpath: \
-	${testpath}
