--- io.openems.edge.fronius.gen24.neu/readme.adoc
@@ -1,27 +1,0 @@
-= SMA
-
-== ESS
-
-Tested on
-- SMA SunnyIsland 6.0H
-- Sunny Island 4.4M
-- SMA STP 10.0 Hybrid ESS. (The implementation is split into Battery, Inverter and DC Charger. The inverter only works with that battery, but the implementation had to be split up as the devices use different Modbus Unit IDs.)
-
-== Meter
-
-Tested on
-- SMA Sunny Home Manager 2.0 Integrated Meter
-
-This implementation uses the integrated energy meter of the SMA Sunny Home Manager 2.0
-Data needs to polled through Modbus from an attached SMA inverter as the HM2.0 does not have a local API on the device itself
-
-Implemented Natures:
-- ElectricityMeter
-
-== PV Inverter
-
-Tested on
-- https://www.sma.de/produkte/solar-wechselrichter/sunny-tripower-80-100.html[SMA SUNNY TRIPOWER 8.0 / 10.0 ]
-- https://www.sma.de/produkte/solar-wechselrichter/sunny-tripower-15000tl-20000tl-25000tl.html[SMA SUNNY TRIPOWER 15000TL / 20000TL / 25000TL]
-
-https://github.com/OpenEMS/openems/tree/develop/io.openems.edge.sma[Source Code icon:github[]]
