--- io.openems.edge.fronius.gen24.neu/src/io/openems/edge/fronius/ess/enums/SetControlMode.java
@@ -1,34 +1,0 @@
-package io.openems.edge.fronius.ess.enums;
-
-import io.openems.common.types.OptionsEnum;
-
-public enum SetControlMode implements OptionsEnum {
-	UNDEFINED(-1, "Undefined"), //
-	DISABLED(0, "Disabled"), //
-	CHARGE_LIMIT(1, "Charge limit"), //
-	DISCHARGE_LIMIT(2, "Discharge limit"), //
-	CHARGE_AND_DISCHARGE_LIMIT(3, "Charge and discharge limit");
-
-	private final int value;
-	private final String name;
-
-	private SetControlMode(int value, String name) {
-		this.value = value;
-		this.name = name;
-	}
-
-	@Override
-	public int getValue() {
-		return this.value;
-	}
-
-	@Override
-	public String getName() {
-		return this.name;
-	}
-
-	@Override
-	public OptionsEnum getUndefined() {
-		return UNDEFINED;
-	}
-}
