--- io.openems.edge.fronius.gen24.neu/test/io/openems/edge/fronius/ess/gen24/batteryinverter/MyConfig.java
@@ -1,75 +1,0 @@
-package io.openems.edge.fronius.ess.gen24.batteryinverter;
-
-import io.openems.common.test.AbstractComponentConfig;
-import io.openems.edge.fronius.ess.gen24.batteryinverter.Config;
-import io.openems.edge.fronius.ess.gen24.batteryinverter.ControlMode;
-
-@SuppressWarnings("all")
-public class MyConfig extends AbstractComponentConfig implements Config {
-
-	protected static class Builder {
-		private String id;
-		private String modbusId;
-		private int modbusUnitId;
-		private ControlMode controlMode;
-
-		private Builder() {
-		}
-
-		public Builder setId(String id) {
-			this.id = id;
-			return this;
-		}
-
-		public Builder setModbusId(String modbusId) {
-			this.modbusId = modbusId;
-			return this;
-		}
-
-		public Builder setModbusUnitId(int modbusUnitId) {
-			this.modbusUnitId = modbusUnitId;
-			return this;
-		}
-
-		public Builder setControlMode(ControlMode controlMode) {
-			this.controlMode = controlMode;
-			return this;
-		}
-
-		public MyConfig build() {
-			return new MyConfig(this);
-		}
-
-	}
-
-	/**
-	 * Create a Config builder.
-	 *
-	 * @return a {@link Builder}
-	 */
-	public static Builder create() {
-		return new Builder();
-	}
-
-	private final Builder builder;
-
-	private MyConfig(Builder builder) {
-		super(Config.class, builder.id);
-		this.builder = builder;
-	}
-
-	@Override
-	public String modbus_id() {
-		return this.builder.modbusId;
-	}
-
-	@Override
-	public int modbusUnitId() {
-		return this.builder.modbusUnitId;
-	}
-
-	@Override
-	public ControlMode controlMode() {
-		return this.builder.controlMode;
-	}
-}
