+++ io.openems.edge.fronius/bnd.bnd
@@ -11,6 +11,12 @@ Bundle-Version: 1.0.0.${tstamp}
 	io.openems.edge.meter.api,\
 	io.openems.edge.pvinverter.api,\
 	io.openems.j2mod,\
+	io.openems.edge.battery.api,\
+	io.openems.edge.batteryinverter.api,\
+	io.openems.edge.ess.api,\
+	io.openems.edge.meter.api,\
+	io.openems.edge.timedata.api,\
+	
 
 -testpath: \
 	${testpath},\
